Flytter tilDto-funksjon fra Vedtaksperiode.kt til VedtaksperiodeDto.kt
