+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/dto/VedtaksperiodeDto.kt
@@ -39,6 +39,30 @@ enum class VedtaksperiodeStatus {
     UENDRET,
 }
 
+fun List<Vedtaksperiode>.tilVedtaksperiodeDto(tidligereVedtaksperioder: List<Vedtaksperiode>?) =
+    map {
+        it.tilDto(tidligereVedtaksperioder?.find { v -> v.id == it.id })
+    }
+
+fun Vedtaksperiode.tilDto(forrigeVedtaksperiode: Vedtaksperiode?) =
+    VedtaksperiodeDto(
+        id = id,
+        fom = fom,
+        tom = tom,
+        målgruppeType = målgruppe,
+        aktivitetType = aktivitet,
+        status = utledStatus(forrigeVedtaksperiode),
+    )
+
+private fun Vedtaksperiode.utledStatus(forrigeVedtaksperiode: Vedtaksperiode?): VedtaksperiodeStatus =
+    when {
+        forrigeVedtaksperiode == null -> VedtaksperiodeStatus.NY
+        this.fom == forrigeVedtaksperiode.fom && this.tom == forrigeVedtaksperiode.tom ->
+            VedtaksperiodeStatus.UENDRET
+
+        else -> VedtaksperiodeStatus.ENDRET
+    }
+
 fun List<VedtaksperiodeDto>.tilDomene() = map { it.tilDomene() }
 
 fun List<VedtaksperiodeDto>.tilVedtaksperiodeBeregning() = map { VedtaksperiodeBeregning(it) }
