+++ src/test/kotlin/no/nav/tilleggsstonader/sak/vedtak/barnetilsyn/beregning/BeregningUtilsStepDefinitons.kt
@@ -17,7 +17,7 @@ import no.nav.tilleggsstonader.sak.vedtak.domain.PeriodeMedDager
 import no.nav.tilleggsstonader.sak.vedtak.domain.Uke
 import no.nav.tilleggsstonader.sak.vedtak.domain.VedtaksperiodeBeregning
 import no.nav.tilleggsstonader.sak.vedtak.domain.VedtaksperiodeBeregningUtil.tilUke
-import no.nav.tilleggsstonader.sak.vedtak.domain.tilVedtaksperiodeDto
+import no.nav.tilleggsstonader.sak.vedtak.dto.tilVedtaksperiodeDto
 import org.assertj.core.api.Assertions.assertThat
 
 @Suppress("ktlint:standard:function-naming", "unused")
